Addresses PR feedback
