+++ host_core/native/hostcore_wasmcloud_native/src/par.rs
@@ -1,10 +1,8 @@
 use crate::{Claims, ProviderArchiveResource};
-use chrono::NaiveDateTime;
 
 use provider_archive::ProviderArchive;
 use rustler::{Env, Error};
 use std::env::temp_dir;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 pub fn on_load(env: Env) -> bool {
     rustler::resource!(ProviderArchiveResource, env);
@@ -26,29 +24,9 @@ pub(crate) fn get_vendor(par: &ProviderArchive) -> Result<String, Error> {
 }
 
 pub(crate) fn extract_claims(par: &ProviderArchive) -> Result<Claims, Error> {
-    match par.claims() {
-        Some(c) => Ok(convert_provider_claims(c)),
-        None => Err(Error::Term(Box::new("No claims found in provider archive"))),
-    }
-}
-
-/// Convert provider claims to our crate claims type
-pub fn convert_provider_claims(
-    c: wascap::jwt::Claims<wascap::jwt::CapabilityProvider>,
-) -> crate::Claims {
-    let metadata = c.metadata.unwrap_or_default();
-    let revision = crate::revision_or_iat(metadata.rev, c.issued_at);
-    crate::Claims {
-        issuer: c.issuer,
-        public_key: c.subject,
-        revision,
-        tags: None,
-        version: metadata.ver,
-        name: metadata.name,
-        expires_human: stamp_to_human(c.expires).unwrap_or_else(|| "never".to_string()),
-        not_before_human: stamp_to_human(c.not_before).unwrap_or_else(|| "immediately".to_string()),
-        ..Default::default()
-    }
+    par.claims()
+        .map(Claims::from)
+        .ok_or_else(|| Error::Term(Box::new("No claims found in provider archive")))
 }
 
 pub(crate) fn extract_target_bytes(par: &ProviderArchive) -> Result<Vec<u8>, Error> {
@@ -97,23 +75,4 @@ fn normalize_for_filename(input: &str) -> String {
 
 fn native_target() -> String {
     format!("{}-{}", std::env::consts::ARCH, std::env::consts::OS)
-}
-
-fn stamp_to_human(stamp: Option<u64>) -> Option<String> {
-    stamp.map(|s| {
-        let now = NaiveDateTime::from_timestamp(since_the_epoch().as_secs() as i64, 0);
-        let then = NaiveDateTime::from_timestamp(s as i64, 0);
-
-        let diff = then - now;
-
-        let ht = chrono_humanize::HumanTime::from(diff);
-        format!("{}", ht)
-    })
-}
-
-fn since_the_epoch() -> Duration {
-    let start = SystemTime::now();
-    start
-        .duration_since(UNIX_EPOCH)
-        .expect("A timey wimey problem has occurred!")
 }
